With transform boxes test moved to test/bbox
