+++ test/geometry/transform/test_projwarp.py
@@ -302,7 +302,7 @@ class TestPerspectiveTransform3D:
     @pytest.mark.parametrize("batch_size", [1, 2])
     def test_get_perspective_transform3d_2(self, batch_size, device, dtype):
         torch.manual_seed(0)
-        src = kornia.bbox_generator3d(
+        src = kornia.bbox_generator_3d(
             torch.randint_like(torch.ones(batch_size), 0, 50, dtype=dtype),
             torch.randint_like(torch.ones(batch_size), 0, 50, dtype=dtype),
             torch.randint_like(torch.ones(batch_size), 0, 50, dtype=dtype),
@@ -310,7 +310,7 @@ class TestPerspectiveTransform3D:
             torch.randint(0, 50, (1,), dtype=dtype).repeat(batch_size),
             torch.randint(0, 50, (1,), dtype=dtype).repeat(batch_size),
         ).to(device=device, dtype=dtype)
-        dst = kornia.bbox_generator3d(
+        dst = kornia.bbox_generator_3d(
             torch.randint_like(torch.ones(batch_size), 0, 50, dtype=dtype),
             torch.randint_like(torch.ones(batch_size), 0, 50, dtype=dtype),
             torch.randint_like(torch.ones(batch_size), 0, 50, dtype=dtype),
